+++ Source/ColonyManagerRedux/Helpers/Resources.cs
@@ -62,7 +62,8 @@ public static class Resources
         Tame = ContentFinder<Texture2D>.Get("UI/Icons/Animal/Tame"),
         Slaughter = ContentFinder<Texture2D>.Get("UI/Icons/Animal/Slaughter"),
         ReleaseToTheWild = ContentFinder<Texture2D>.Get("UI/Designators/ReleaseToTheWild"),
-        ManagerTab_Gizmo = ContentFinder<Texture2D>.Get("UI/Commands/CMR_ManagerTab_Gizmo");
+        ManagerTab_Gizmo = ContentFinder<Texture2D>.Get("UI/Commands/CMR_ManagerTab_Gizmo"),
+        Venerated = ContentFinder<Texture2D>.Get("UI/Icons/CMR_venerated");
 
     public static Texture2D GetLifeStageIcon(int lifeStageIndex)
     {